feat(pando cli): split up setup tests via describe
