+++ packages/pando/src/setup/step1/prompts.mts
@@ -2,10 +2,10 @@ import { PMOptions } from 'shared/types.mts'
 
 export const step1Msg = 'Step 1: Confirm your package manager'
 
-export function confirmDetectedPm(pm: PMOptions) {
-  return `We've detected that your package manager is ${pm}. Does that sound right?`
+export function confirmDetectedPm(lockFileName: string, pm: PMOptions) {
+  return `We've detected a ${lockFileName} file. Would you like us to install dependencies using ${pm}?`
 }
 
-export function pmNameConfirmed(pm: PMOptions) {
-  return `we've determined that ${pm} is your package manager! Great.`
-}
+// export function pmNameConfirmed(pm: PMOptions, packages: string[]) {
+//   return `Installing ${packages.join(', ')} using ${pm}...`
+// }
